requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
decorator==5.1.1
Django==5.0.7
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
executing==2.0.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
msgpack==1.0.8
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg==3.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
pytest==8.3.2
redis==5.0.8
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
wcwidth==0.2.13
zope.interface==7.0.3
